third_party/WebKit/LayoutTests/inspector/elements/styles-2/inject-stylesheet-expected.txt

CONSOLE MESSAGE: line 10: iframe loaded
Tests that injected user stylesheets are reflected in the Styles pane.


Main frame style:
background-attachment: scroll
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:3
background-clip: border-box
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:3
background-color: rgb(0, 0, 255)
    blue - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:3
background-image: none
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:3
background-origin: padding-box
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:3
background-size: auto
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:3
color: rgb(255, 0, 0)
    red - #main injected stylesheet
display: block
    block - div user agent stylesheet
-webkit-border-image: url("data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElE…s+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==") 100% / 1 / 0px stretch
    url("data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElE…s+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==") - #main injected stylesheet
[expanded] 
element.style  { ()

[expanded] 
media="screen" (inject-stylesheet.html)
#main  { (inject-stylesheet.html:5 -> inject-stylesheet.html:5:3)
    background: blue;
        background-image: initial;
        background-position-x: initial;
        background-position-y: initial;
        background-size: initial;
        background-repeat-x: initial;
        background-repeat-y: initial;
        background-attachment: initial;
        background-origin: initial;
        background-clip: initial;
        background-color: blue;

[expanded] 
#main  { (injected stylesheet)
    color: red;
    -webkit-border-image: url("data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElE…s+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==");

[expanded] 
div  { (user agent stylesheet)
    display: block;

iframe style:
background-attachment: scroll
    initial - #iframebody injected stylesheet
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3
background-clip: border-box
    initial - #iframebody injected stylesheet
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3
background-color: rgb(255, 0, 0)
    red - #iframebody injected stylesheet
    OVERLOADED green - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3
background-image: none
    initial - #iframebody injected stylesheet
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3
background-origin: padding-box
    initial - #iframebody injected stylesheet
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3
background-size: auto
    initial - #iframebody injected stylesheet
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3
display: block
    block - body user agent stylesheet
margin-bottom: 8px
    8px - body user agent stylesheet
margin-left: 8px
    8px - body user agent stylesheet
margin-right: 8px
    8px - body user agent stylesheet
margin-top: 8px
    8px - body user agent stylesheet
[expanded] 
element.style  { ()

[expanded] 
#iframebody  { (injected stylesheet)
    background: red;
        background-image: initial;
        background-position-x: initial;
        background-position-y: initial;
        background-size: initial;
        background-repeat-x: initial;
        background-repeat-y: initial;
        background-attachment: initial;
        background-origin: initial;
        background-clip: initial;
        background-color: red;

[expanded] 
media="screen" (inject-stylesheet-iframe-data.html)
body  { (inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3)
/-- overloaded --/     background: green;
    /-- overloaded --/     background-image: initial;
    /-- overloaded --/     background-position-x: initial;
    /-- overloaded --/     background-position-y: initial;
    /-- overloaded --/     background-size: initial;
    /-- overloaded --/     background-repeat-x: initial;
    /-- overloaded --/     background-repeat-y: initial;
    /-- overloaded --/     background-attachment: initial;
    /-- overloaded --/     background-origin: initial;
    /-- overloaded --/     background-clip: initial;
    /-- overloaded --/     background-color: green;

[expanded] 
body  { (user agent stylesheet)
    display: block;
    margin: 8px;
        margin-top: 8px;
        margin-right: 8px;
        margin-bottom: 8px;
        margin-left: 8px;
